pyproject.toml
[tool.poetry]
name = "billing"
version = "0.1.0"
description = ""
authors = ["Alisson <alissonfilipe937@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
grpcio = "^1.60.1"
grpcio-tools = "^1.60.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
